Refactor editor game capture tooling
