+++ Assets/Editor/Capture.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1ce7364a2dc44cf3bc5e7a9a2924a8b4
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
