+++ Assets/Editor/Capture/GameRecordingControlsWindow.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: ce4961d3251e44f8a97122235d143c87
