+++ Assets/Editor/Capture/GameRecordingSession.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: aa433d5c69e34d708adaa33631a66178
